+++ src/photometer.cpp
