+++ src/photometer.h
